Request validering ut av Service
